Update stickyNav.js
Added 'padding' class to the next element so that media queries can also be used in conjunction to remove the padding buff when the 'fixed' class is being ignored via media queries. Also added a positional check to init() so the element doesn't disappear on window resize any more. Window resize destroys and inits plugin on horizontal resize as well to accommodate the switch to media queries